requirements.txt
# Core web & API
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
upstash-redis
qdrant-client==1.5.0

# Database
sqlalchemy>=2.0.20,<2.1
psycopg2-binary>=2.9,<3
redis>=5.3.0,<6
celery[redis]>=5.3,<6

# ML & Data
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.2.2
joblib==1.3.2
torch>=2.1.0,<2.2.0
transformers>=4.40,<5
sentence-transformers>=2.2,<3
tensorflow==2.14.0
tensorflow-recommenders==0.7.3
pillow>=10.0,<11.0
faker==18.9.0
nltk
spacy
keybert